requirements.txt
Werkzeug==2.2.3
Flask==2.2.3
python-telegram-bot==20.3
uvicorn==0.22.0
asgiref==3.6.0
gunicorn==23.0.0